add_logs_to_python_app_otel_sdk/requirements.txt
Deprecated==1.2.15
importlib_metadata==8.5.0
Nuitka==1.5.1
opentelemetry-api==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-logging==0.50b0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
ordered-set==4.1.0
packaging==24.2
typing_extensions==4.12.2
urllib3==1.26.15
wrapt==1.17.2
zipp==3.21.0
zstandard==0.20.0
